Few Versions in already.
Currently everything is working fine.
There is no point system.

If you try rotate a piece and it collides with something, it will not
rotate. In a future version the piece *May* try to move into a place
that does not collide instead of just not rotating.

Input is buggy at best, this is the next task.

At this point the main focus is using this as a learning tool for
architecture things and best practices.

Things learned:
Do not clutter the main loop.
Use getter and setter methods.
Functions should only do ONE defined task.

In previous version, the active and stored pieces weren't part of the
board, they were separate. This caused a lot of collision logic to be in
the main loop.

Now active and stored are part of the board, and  it does collision
detection each time the player tries to move the piece.