Input Handler
Added the input handler module, which acts like a singleton.
It is a layer between the actual input and the resulting action in the
game.

It uses functions as first class arguments in a list, therefore allowing
key reconfiguration later on, if I want to make that part of the game.
(I don't right now.)

The input feels much more controlled.
Actions take place immediately and the delay for repeated actions can be
changed in the InputHandler file.