Game States + Line Count
Added in game states and cleaned up some unused things.
Also added a display of the number of lines cleared.
This involved initializing the font module in the board module.

List of next pieces will not come until I start fiddling with the way
the whole game looks, because I need a place to display it.