Levelling system
Added basic levelling system where the fall rate of the block =
2+lines_cleared/10

Also enabled the active piece to "slide" a little bit before it sticks
to the board. This also changes as the lines cleared goes up, because
the piece is falling faster and therefore "sticking" more quickly.

Fixed bug where the stored piece would disappear if player tried to
switch at the wrong time.