Re-factoring and Tweaks
Moved things into separate files, giving the tetromino class it's own
file, constants their own file, and moved the game board into a Board.py
file.

Also added a delay to the repeated movement in the x direction in the
input handler.

The random generator module ensures the distribution of the tetrominos
is the same as that of the official game.
It does this by generating seven at time that are all different.