User profiles, Point system. Fixed Large Bug.
Added a UserProfiles Module because all my friends play with a different
key configuration. Now only one line needs to be changed depending on
which one is playing.

A point system similar to but slightly simpler than the one in
TetrisFriends has been implemented.

While researching that I found that I've mis-spelled Tetrimino The
entire game. I have no plans on fixing this.

There was a large error with he way the board cleared lines that was
giving "almost" right behaviour until now, hence why it went un-noticed.
It became obvious when the coding the point system.

Also realized the board only needs to try to clear lines when a piece
has been absorbed by the board, not every frame.

Decided that the board should only render the actual gameboard, no
scored or information. The replacement has not been implemented yet but
it will be in a render function in main.

Still to come: Revamped display, rotate counter-clockwise.