Main Menu
Created the Menu and button files which make the main menu for the game.
This required adding a reject button to the input handler, and a new
EXIT gamestate.

Game states have been moved to the constants file, so that other modules
can use them (menu buttons have attributes which are the game states.)