Next pieces
The game will now show the next five pieces coming along the righthand
side of the board. This was implemented with a deque, which turned out
to be overkill. All I wanted was a queue that could be iterated over
without removing the contents (for rendering purposes). Turns out the
standard list can do this but I had already written the code to use the
deque.

Why is python's queue implementation less functional than it's list?
Maybe lists are just awesome.